requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
python-docx
python-dotenv
langchain
openai
tiktoken
